+++ tests/unit/TestGsmSwapFuzz.t.sol
@@ -255,7 +255,7 @@ contract TestGsmSwapFuzz is TestGhoBase {
     Gsm gsm = new Gsm(address(GHO_TOKEN), address(newToken), address(newPriceStrategy));
     gsm.initialize(address(this), TREASURY, type(uint128).max, address(GHO_RESERVE));
     GHO_RESERVE.addEntity(address(gsm));
-    GHO_RESERVE.setLimit(address(gsm), type(uint256).max);
+    GHO_RESERVE.setLimit(address(gsm), type(uint128).max);
     deal(address(GHO_TOKEN), address(GHO_RESERVE), type(uint256).max);
 
     if (buyFeeBps > 0 || sellFeeBps > 0) {
@@ -307,7 +307,7 @@ contract TestGsmSwapFuzz is TestGhoBase {
     Gsm gsm = new Gsm(address(GHO_TOKEN), address(newToken), address(newPriceStrategy));
     gsm.initialize(address(this), TREASURY, type(uint128).max, address(GHO_RESERVE));
     GHO_RESERVE.addEntity(address(gsm));
-    GHO_RESERVE.setLimit(address(gsm), type(uint256).max);
+    GHO_RESERVE.setLimit(address(gsm), type(uint128).max);
 
     if (buyFeeBps > 0 || sellFeeBps > 0) {
       FixedFeeStrategy newFeeStrategy = new FixedFeeStrategy(buyFeeBps, sellFeeBps);
@@ -352,7 +352,7 @@ contract TestGsmSwapFuzz is TestGhoBase {
     Gsm gsm = new Gsm(address(GHO_TOKEN), address(newToken), address(newPriceStrategy));
     gsm.initialize(address(this), TREASURY, type(uint128).max, address(GHO_RESERVE));
     GHO_RESERVE.addEntity(address(gsm));
-    GHO_RESERVE.setLimit(address(gsm), type(uint256).max);
+    GHO_RESERVE.setLimit(address(gsm), type(uint128).max);
 
     if (buyFeeBps > 0 || sellFeeBps > 0) {
       FixedFeeStrategy newFeeStrategy = new FixedFeeStrategy(buyFeeBps, sellFeeBps);
@@ -397,7 +397,7 @@ contract TestGsmSwapFuzz is TestGhoBase {
     Gsm gsm = new Gsm(address(GHO_TOKEN), address(newToken), address(newPriceStrategy));
     gsm.initialize(address(this), TREASURY, type(uint128).max, address(GHO_RESERVE));
     GHO_RESERVE.addEntity(address(gsm));
-    GHO_RESERVE.setLimit(address(gsm), type(uint256).max);
+    GHO_RESERVE.setLimit(address(gsm), type(uint128).max);
 
     if (buyFeeBps > 0 || sellFeeBps > 0) {
       FixedFeeStrategy newFeeStrategy = new FixedFeeStrategy(buyFeeBps, sellFeeBps);
@@ -627,7 +627,7 @@ contract TestGsmSwapFuzz is TestGhoBase {
     Gsm gsm = new Gsm(address(GHO_TOKEN), address(newToken), address(newPriceStrategy));
     gsm.initialize(address(this), TREASURY, uint128(assetAmount), address(GHO_RESERVE));
     GHO_RESERVE.addEntity(address(gsm));
-    GHO_RESERVE.setLimit(address(gsm), type(uint256).max);
+    GHO_RESERVE.setLimit(address(gsm), type(uint128).max);
 
     if (buyFeeBps > 0 || sellFeeBps > 0) {
       FixedFeeStrategy newFeeStrategy = new FixedFeeStrategy(buyFeeBps, sellFeeBps);
@@ -726,7 +726,7 @@ contract TestGsmSwapFuzz is TestGhoBase {
     Gsm gsm = new Gsm(address(GHO_TOKEN), address(newToken), address(newPriceStrategy));
     gsm.initialize(address(this), TREASURY, type(uint128).max, address(GHO_RESERVE));
     GHO_RESERVE.addEntity(address(gsm));
-    GHO_RESERVE.setLimit(address(gsm), type(uint256).max);
+    GHO_RESERVE.setLimit(address(gsm), type(uint128).max);
 
     if (buyFeeBps > 0 || sellFeeBps > 0) {
       FixedFeeStrategy newFeeStrategy = new FixedFeeStrategy(buyFeeBps, sellFeeBps);
